Adds Chirila Solo Encounter.
Player can start, escape, and fail to escape encounters with a single
chirilla in the new encounter structure.
Breaking an encounter when one of multiple chirillas falls to 0 stam
has not been tested.
Enemies are now stunned (temp sprite) when encounters end.
Adds stamina bar and value to chirilla for testing purposes.
Player is now undetected in grass (tested).